Use chroot for Zookeper in Kafka settings

+++ containers/build.sh
@@ -3,11 +3,17 @@ containers_dir="${BASH_SOURCE%/*}"
 if [[ ! -d "$DIR" ]]; then DIR="$PWD"; fi
 source "$containers_dir/../parse-env.sh"
 
-opts=$2
+container_version=$version
+subver=$2
+opts=$3
 
 echo 'Contrail version: '$version
 echo 'Contrail registry: '$registry
 echo 'Contrail repository: '$repository
+if [ -n "$subver" ]; then
+  container_version=$version'-'$subver
+  echo 'Container version: '$container_version
+fi
 if [ -n "$opts" ]; then
   echo 'Options: '$opts
 fi
@@ -26,17 +32,20 @@ build () {
   local container_name='contrail'${container_name}
   echo 'Building '$container_name
   local logfile='build-'$container_name'.log'
-  docker build -t ${registry}'/'${container_name}:${version} \
+  tags="-t ${registry}/${container_name}:${container_version}"
+  if [ -n $subver ]; then
+    tags+=" -t ${registry}/${container_name}:${version}"
+  fi
+  failed=0
+  docker build ${tags} \
     --build-arg CONTRAIL_VERSION=${version} \
     --build-arg CONTRAIL_REGISTRY=${registry} \
     --build-arg REPOSITORY=${repository} \
     ${opts} $dir |& tee $logfile
-  if [ ${PIPESTATUS[0]} -eq 0 ]; then
-    docker push ${registry}'/'${container_name}:${version} |& tee -a $logfile
-    if [ ${PIPESTATUS[0]} -eq 0 ]; then
-      rm $logfile
-    fi
-  fi
+  state=${PIPESTATUS[0]}
+  test $state -eq 0 && docker push ${registry}'/'${container_name}:${container_version} |& tee -a $logfile && state=${PIPESTATUS[0]}
+  test $state -eq 0 && docker push ${registry}'/'${container_name}:${version} |& tee -a $logfile && state=${PIPESTATUS[0]}
+  test $state -eq 0 && rm $logfile
   if [ -f $logfile ]; then
     was_errors=1
   fi
